app/requirements.txt
numpy==1.21.5
opencv_python_headless
Pillow==9.4.0
requests==2.25.1
streamlit==1.20.0
tensorflow==2.10.0
tensorflow_hub==0.12.0